feat(feed): campaign container component for fetching data
