+++ client/src/components/CampaignContainer/CampaignContainer.jsx
@@ -3,8 +3,7 @@ import {
     StyleSheet,
     View,
 } from "react-native";
-
-import CapmpaignCard from "../CampaignCard/CampaignCard";
+import CampaignsComponent from "../CampaignsComponent/CampaignsComponent";
 
 const campaignExamples = [
     {
@@ -103,44 +102,11 @@ const campaignExamples = [
 
 const CampaignContainer = (props) => {
 
-    const renderItem = ({ item }) => (
-        <CapmpaignCard
-            status={item.status}
-            title={item.titulo}
-            user={item.userId}
-            location={item.ubicacion}
-            startDate={item.fechaInicio}
-            finishDate={item.fechaExpiracion}
-            progress={Math.round(item.donativosTotales * 100 / item.metaDonativos)}
-            categoriaEnseres={item.categoriaEnseres}
-            categoriaFrutasVerduras={item.categoriaFrutasVerduras}
-            categoriaNoPerecederos={item.categoriaNoPerecederos}
-            categoriaRopa={item.categoriaRopa}
-        />
-    )
-
     return (
-        <View style={Styles.container}>
-
-            <FlatList
-                data={campaignExamples}
-                renderItem={renderItem}
-                keyExtractor={item => item.campaignId}
-                showsVerticalScrollIndicator={false}
-            />
-
-        </View>
+        <CampaignsComponent
+            campaigns={campaignExamples}
+        />
     );
 };
-
-const Styles = StyleSheet.create({
-    container: {
-        flex: 9,
-        backgroundColor: "white",
-        paddingTop: 10,
-        paddingRight: 20,
-        paddingLeft: 20,
-    }
-})
 
 export default CampaignContainer;
